FP-615 Добавлены файлы процесса для создания и экспорта реестра окончания раб дня

+++ trunk/payments/tests/java/org/flexpay/payments/util/registries/TestEndOperationDayRegistryGenerator.java
@@ -5,6 +5,8 @@ import org.flexpay.common.persistence.registry.Registry;
 import org.flexpay.common.service.RegistryService;
 import org.flexpay.orgs.persistence.Organization;
 import org.flexpay.orgs.persistence.PaymentPoint;
+import org.flexpay.orgs.service.OrganizationService;
+import org.flexpay.orgs.service.PaymentPointService;
 import org.flexpay.payments.test.PaymentsSpringBeanAwareTestCase;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +20,10 @@ public class TestEndOperationDayRegistryGenerator extends PaymentsSpringBeanAwar
 	@Autowired
 	EndOperationDayRegistryGenerator endOperationDayRegistryGenerator;
 	@Autowired
+	PaymentPointService paymentPointService;
+	@Autowired
+	OrganizationService organizationService;
+	@Autowired
 	@Qualifier("registryService")
 	RegistryService registryService;
 
@@ -29,7 +35,20 @@ public class TestEndOperationDayRegistryGenerator extends PaymentsSpringBeanAwar
 		Date beginDate = df.parse("2009-06-07 00:00:00");
 		Date endDate = df.parse("2009-06-07 23:59:59");
 
-		Registry registry = endOperationDayRegistryGenerator.generate(new Stub<PaymentPoint>(2L), new Stub<Organization>(4L), beginDate, endDate);
+		PaymentPoint paymentPoint = paymentPointService.read(new Stub<PaymentPoint>(2L));
+		if (paymentPoint == null) {
+			log.error("Payment point with id - {} does not exist", 2L);
+			return;
+		}
+		log.debug("Found paymentPoint - {}", paymentPoint);
+
+		Organization organization = organizationService.readFull(new Stub<Organization>(4L));
+		if (organization == null) {
+			log.error("Organization with id - {} does not exist", 4L);
+			return;
+		}
+
+		Registry registry = endOperationDayRegistryGenerator.generate(paymentPoint, organization, beginDate, endDate);
 
 		registryService.deleteRecords(Stub.stub(registry));
 		registryService.delete(registry);
